Split the optuna from the other main file
